macho: implement basic section movement and reallocation

+++ src/link/MachO/commands.zig
@@ -267,7 +267,7 @@ pub const SegmentCommand = struct {
         return null;
     }
 
-    pub fn findFreeSpace(self: SegmentCommand, object_size: u64, min_alignment: u32, start: ?u64) u64 {
+    pub fn findFreeSpace(self: SegmentCommand, object_size: u64, min_alignment: u64, start: ?u64) u64 {
         var offset: u64 = if (start) |v| v else self.inner.fileoff;
         while (self.detectAllocCollision(offset, object_size)) |item_end| {
             offset = mem.alignForwardGeneric(u64, item_end, min_alignment);
